datos descendentes en querys de reportes y status actualizado a 0

+++ src/main/java/com/consiti/entity/Reportes.java
@@ -7,7 +7,7 @@ public interface Reportes {
 
 	 String getEntity();
 	 String getKey();
-	 Integer getValue();
-	 Integer getKpi();
+	 Double getValue();
+	 Double getKpi();
 	 Double getProgress();
 }
